Modify cond_reltimedwait with uint64 argument to support long time wait

+++ wamr-sdk/app/libc-builtin-sysroot/include/pthread.h
@@ -6,6 +6,8 @@
 #ifndef _WAMR_LIB_PTHREAD_H
 #define _WAMR_LIB_PTHREAD_H
 
+#include <stdint.h>
+
 /* Data type define of pthread, mutex, cond and key */
 typedef unsigned int pthread_t;
 typedef unsigned int pthread_mutex_t;
@@ -41,7 +43,7 @@ int pthread_cond_init(pthread_cond_t *cond, const void *attr);
 int pthread_cond_wait(pthread_cond_t *cond, pthread_mutex_t *mutex);
 
 int pthread_cond_timedwait(pthread_cond_t *cond, pthread_mutex_t *mutex,
-                           unsigned int useconds);
+                           uint64_t useconds);
 
 int pthread_cond_signal(pthread_cond_t *cond);
